gui changes
