+++ .gitignore
@@ -1,0 +1,3 @@
+/build
+/dist
+/__pycache__
