--- excel_config.json
@@ -1,10 +1,0 @@
-{
-    "workbook": "C:\\excel_exports\\Active_Employees_for_Sarah.xlsx",
-    "file": "C:\\excel_exports\\Active_Employees_for_Sarah.sql",
-    "rows": 205,
-    "columns": 3,
-    "row-start" : 2,
-    "col-start": 1,
-    "delimiter": "%",
-    "program_type": "insert"
-}
